LowLevelFEM.jl

Documentation for LowLevelFEM.jl

Functions

LowLevelFEM.Problem — Type

Problem(names; thickness=..., type=..., bandwidth=...)

A structure containing the most important data of the problem.

name of the model (in gmsh)

type of the problem: 3D "Solid", "PlaneStrain", "PlaneStress" or "AxiSymmetric" In the case of "AxiSymmetric", the axis of symmetry is the "y" axis, while the geometry must be drawn in the positive "x" half-plane.

bandwidth optimization using built-in gmsh function. Possibilities: "RCMK" (default), "Hilbert", "Metis" or "none"

dimension of the problem, determined from type

material constants: Physical group, Young's modulus, Poisson's ratio, mass density (in vector of tuples names)

thickness of the plate

number of nodes (non)

Types:

names: Vector{Touple{String, Float64, Float64, Float64}}

type: String

bandwidth: String

dim: Integer

thickness: Float64

non: Integer

LowLevelFEM.StressField — Type

StressField(sigma, numElem, nsteps)

A structure containing the data of a stress or strain field.

sigma: vector of ElementNodeData type stress data (see gmsh.jl)

numElem: vector of tags of elements

nsteps: number of stress fields stored in sigma (for animations).

Types:

sigma: Vector{Matrix{Float64}}

numElem: Vector{Integer}

nsteps: Integer

LowLevelFEM.CDM — Method

FEM.CDM(K, M, C, f, u0, v0, T, Δt)

Solves a transient dynamic problem using central difference method (CDM) (explicit). K is the stiffness Matrix, M is the mass matrix, C is the damping matrix, f is the load vector, u0 is the initial displacement, v0 is the initial velocity, T is the upper bound ot the time intervall (lower bound is zero) and Δt is the time step size. Returns the displacement vectors and velocity vectors in each time step arranged in the columns of the two matrices u and v and a vector t of the time instants used.

Return: u, v, t

Types:

K: SparseMatrix

M: SparseMatrix

C: SparseMatrix

f: Vector{Float64}

u0: Vector{Float64}

v0: Vector{Float64}

T: Float64

Δt: Float64

u: Matrix{Float64}

v: Matrix{Float64}

t: Vector{Float64}

LowLevelFEM.CDMaccuracyAnalysis — Method

FEM.CDMaccuracyAnalysis(ωₘᵢₙ, ωₘₐₓ, Δt, type; n=100, α=..., ξ=..., β=..., show_β=..., show_ξ=...)

Gives some functions (graphs) for accuracy analysis of the CDM method. ωₘᵢₙ and ωₘₐₓ are the square root of smallest and largest eigenvalues of the Kϕ=ω²Mϕ eigenvalue problem, Δt is the time step size. type is one of the following values:

"SR": spectral radius

"PDR": physical damping ratio

"ADR": algorithmic damping ratio

"PE": period error

For details see [3]. n is the number of points in the graph. The damping matrix is assembled in the following ways: C=αM+βK or C=αM+β₁K+β₂KM⁻¹K+β₃KM⁻¹KM⁻¹K+⋅⋅⋅. The latter corresponds to the damping characteristic characterized by a power series consisting of powers of the natural frequencies with odd exponents. ξᵢ (ξ in the argument list) are the values ​​of the individual members of the series corresponding to the ωₘₐₓ value. βᵢ (β in the argument list) are the coefficients of the series. (see [4]) Either ξ or β must be specified. ξ or β are scalars or vectors. If show_β or show_ξ is true, the corresponding β or ξ values will be sent to the output. Returns a tuple of x and y values of the graph. (Can be plotted with plot(xy))

Return: xy

Types:

K: SparseMatrix

M: SparseMatrix

ωₘᵢₙ: Float64

ωₘₐₓ: Float64

Δt: Float64

n: Int64

α: Float64

β: Float64 of Vector{Float64}

ξ: Float64 of Vector{Float64}

show_β: Boolean

show_ξ: Boolean

xy: Tuple{Vector{Float64},Vector{Float64}}

LowLevelFEM.HHT — Method

FEM.HHT(K, M, f, u0, v0, T, Δt; α=..., δ=..., γ=..., β=...)

Solves a transient dynamic problem using HHT-α method[1] (implicit). K is the stiffness Matrix, M is the mass matrix, f is the load vector, u0 is the initial displacement, v0 is the initial velocity, T is the upper bound ot the time intervall (lower bound is zero) and Δt is the time step size. Returns the displacement vectors and velocity vectors in each time step arranged in the columns of the two matrices u and v and a vector t of the time instants used. For the meaning of α, β and γ see [1]. If δ is given, γ=0.5+δ and β=0.25⋅(0.5+γ)².

Return: u, v, t

Types:

K: SparseMatrix

M: SparseMatrix

f: Vector{Float64}

u0: Vector{Float64}

v0: Vector{Float64}

T: Float64

Δt: Float64

α: Float64

β: Float64

γ: Float64

δ: Float64

u: Matrix{Float64}

v: Matrix{Float64}

t: Vector{Float64}

LowLevelFEM.HHTaccuracyAnalysis — Method

FEM.HHTaccuracyAnalysis(ωₘᵢₙ, ωₘₐₓ, Δt, type; n=100, α=0.0, δ=0.0, γ=0.5 + δ, β=0.25 * (0.5 + γ)^2)

Gives some functions (graphs) for accuracy analysis of the HHT-α method[1]. ωₘᵢₙ and ωₘₐₓ are the square root of smallest and largest eigenvalues of the Kϕ=ω²Mϕ eigenvalue problem, Δt is the time step size. type is one of the following values:

"SR": spectral radius

"ADR": algorithmic damping ratio

"PE": period error

For details see [2] and [3]. n is the number of points in the graph. For the meaning of α, β and γ see [1]. If δ is given, γ=0.5+δ and β=0.25⋅(0.5+γ)². Returns a tuple of x and y values of the graph. (Can be plotted with plot(xy))

Return: xy

Types:

K: SparseMatrix

M: SparseMatrix

ωₘᵢₙ: Float64

ωₘₐₓ: Float64

Δt: Float64

n: Int64

α: Float64

β: Float64

γ: Float64

δ: Float64

xy: Tuple{Vector{Float64},Vector{Float64}}

LowLevelFEM.applyBoundaryConditions! — Method

FEM.applyBoundaryConditions!(problem, stiffMat, loadVec, supports)

Applies displacement boundary conditions supports on a stiffness matrix stiffMat and load vector loadVec. Mesh details are in problem. supports is a tuple of name of physical group and prescribed displacements ux, uy and uz.

Return: none

Types:

problem: Problem

stiffMat: SparseMatrix

loadVec: Vector

supports: Vector{Tuple{String, Float64, Float64, Float64}}

LowLevelFEM.applyBoundaryConditions! — Method

FEM.applyBoundaryConditions!(problem, stiffMat, massMat, dampMat, loadVec, supports)

Applies displacement boundary conditions supports on a stiffness matrix stiffMat, mass matrix massMat, damping matrix dampMat and load vector loadVec. Mesh details are in problem. supports is a tuple of name of physical group and prescribed displacements ux, uy and uz.

Return: none

Types:

problem: Problem

stiffMat: SparseMatrix

massMat: SparseMatrix

dampMat: SparseMatrix

loadVec: Vector{Float64}

supports: Vector{Tuple{String, Float64, Float64, Float64}}

LowLevelFEM.applyBoundaryConditions — Method

FEM.applyBoundaryConditions(problem, stiffMat, loadVec, supports)

Applies displacement boundary conditions supports on a stiffness matrix stiffMat and load vector loadVec. Mesh details are in problem. supports is a tuple of name of physical group and prescribed displacements ux, uy and uz. Creates a new stiffness matrix and load vector.

Return: stiffMat, loadVec

Types:

problem: Problem

stiffMat: SparseMatrix

loadVec: Vector

supports: Vector{Tuple{String, Float64, Float64, Float64}}

LowLevelFEM.dampingMatrix — Method

FEM.dampingMatrix(K, M, ωₘₐₓ; α=0.0, ξ=..., β=...)

Generates the damping matrix for proportional damping case. C=αM+βK or C=αM+β₁K+β₂KM⁻¹K+β₃KM⁻¹KM⁻¹K+⋅⋅⋅. The latter corresponds to the damping characteristic characterized by a power series consisting of powers of the natural frequencies with odd exponents. ξᵢ (ξ in the argument list) are the values ​​of the individual members of the series corresponding to the ωₘₐₓ value. βᵢ (β in the argument list) are the coefficients of the series. (see [4]) Either ξ or β must be specified. ξ or β are scalars or vectors. K is the stiffness matrix, M is the mass matrix and ωₘₐₓ is the largest natural frequency.

Return: dampingMatrix

Types:

K: SparseMatrix

M: SparseMatrix

ωₘₐₓ: Float64

α: Float64

ξ: Float64 of Vector{Float64}

β: Float64 of Vector{Float64}

dampingMatrix: SparseMatrix

LowLevelFEM.displacementConstraint — Method

FEM.displacementConstraint(name; ux=..., uy=..., uz=...)

Gives the displacement constraints on name physical group. At least one ux, uy or uz value have to be given (depending on the dimension of the problem). ux, uy or uz can be a constant value, or a function of x, y and z. (E.g. fn(x,y,z)=5*(5-x)); FEM.displacementConstraint("support1", ux=fn)) Return: none

Types:

name: string

ux: Float64 of Function

uy: Float64 of Function

uz: Float64 of Function

LowLevelFEM.generateMesh — Method

FEM.generateMesh(problem, surf, elemSize; approxOrder=1, algorithm=6, quadrangle=0, internalNodes=0)

Obsolate, use gmsh script (.geo) instead.

Return: none

Types:

``: x

LowLevelFEM.getTagForPhysicalName — Method

FEM.getTagForPhysicalName(name)

Returns tags of elements of physical group name.

Return: tags

Types:

name: String

tags: Vector{Integer}

LowLevelFEM.initialDisplacement! — Method

FEM.initialDisplacement!(problem, name, u0; ux=..., uy=..., uz=...)

Changes the displacement values ux, uy and uz (depending on the dimension of the problem) at nodes belonging to physical group name. Original values are in displacement vector u0.

Return: none

Types:

problem: Problem

name: String

u0: Vector{Float64}

ux: Float64

uy: Float64

uz: Float64

LowLevelFEM.initialVelocity! — Method

FEM.initialVelocity!(problem, name, v0; vx=..., vy=..., vz=...)

Changes the velocity values vx, vy and vz (depending on the dimension of the problem) at nodes belonging to physical group name. Original values are in velocity vector v0.

Return: none

Types:

problem: Problem

name: String

v0: Vector{Float64}

vx: Float64

vy: Float64

vz: Float64

LowLevelFEM.largestPeriodTime — Method

FEM.largestPeriodTime(K, M)

Solves the largest period of time for a dynamic problem given by stiffness matrix K and the mass matrix M.`

Return: Δt

Types:

K: SparseMatrix

M: SparseMatrix

Δt: Float64

LowLevelFEM.load — Method

FEM.load(name; fx=..., fy=..., fz=...)

Gives the intensity of distributed load on name physical group. At least one fx, fy or fz value have to be given (depending on the dimension of the problem). fx, fy or fz can be a constant value, or a function of x, y and z. (E.g. fn(x,y,z)=5*(5-x)); FEM.load("load1", fx=fn))

Return: none

Types:

name: string

fx: Float64 of Function

fy: Float64 of Function

fz: Float64 of Function

LowLevelFEM.loadVector — Method

FEM.loadVector(problem, loads)

Solves a load vector of problem. loads is a tuple of name of physical group name, coordinates fx, fy and fz of the intensity of distributed force. It can solve traction or body force depending on the problem. In case of 2D problems and Point physical group means concentrated force. In case of 2D problems and Line physical group means surface force. In case of 2D problems and Surface physical group means body force. In case of 3D problems and Point physical group means concentrated force. In case of 3D problems and Line physical group means edge force. In case of 3D problems and Surface physical group means surface force. In case of 3D problems and Volume physical group means body force.

Return: loadVec

Types:

problem: Problem

loads: Vector{Tuple{String, Float64, Float64, Float64}}

loadVec: Vector

LowLevelFEM.massMatrix — Method

FEM.massMatrix(problem; lumped=...)

Solves the mass matrix of the problem. If lumped is true, solves lumped mass matrix.

Return: massMat

Types:

problem: Problem

lumped: Boolean

massMat: SparseMatrix

LowLevelFEM.material — Method

FEM.material(name; E=2.0e5, ν=0.3, ρ=7.85e-9)

Returns a tuple in which name is the name of a physical group, E is the modulus of elasticity, ν Poisson's ratio and ρ is the mass density.

Return: mat

Types:

mat: Tuple(String, Float64, Float64, Float64)

name: String

E: Float64

ν: Float64

ρ: Float64

LowLevelFEM.nodalAcceleration! — Method

FEM.nodalAcceleration!(problem, name, a0; ax=..., ay=..., az=...)

Changes the acceleration values ax, ay and az (depending on the dimension of the problem) at nodes belonging to physical group name. Original values are in acceleration vector a0.

Return: none

Types:

problem: Problem

name: String

a0: Vector{Float64}

ax: Float64

ay: Float64

az: Float64

LowLevelFEM.nodalForce! — Method

FEM.nodalForce!(problem, name, f0; fx=..., fy=..., fz=...)

Changes the force values fx, fy and fz (depending on the dimension of the problem) at nodes belonging to physical group name. Original values are in load vector f0.

Return: none

Types:

problem: Problem

name: String

f0: Vector{Float64}

fx: Float64

fy: Float64

fz: Float64

LowLevelFEM.plotOnPath — Method

FEM.plotOnPath(problem, pathName, field, points; step=..., plot=..., name=..., visible=...)

Load a 2D plot on a path into a View in gmsh. field is the number of View in gmsh from which the data of a field is imported. pathName is the name of a physical group which contains a curve. The curve is devided into equal length intervals with number of points points. The field is shown at this points. step is the sequence number of displayed step. If no step is given, shows all the aviable steps as an animation. If plot is true, additional return parameter, a tuple of vectors is given back, in which x is a vector of values in horizontal axis, y is a vector of values in vertical axis of a plot (see Plots package). name is the title of graph and visible is a true or false value to toggle on or off the initial visibility in gmsh. This function returns the tag of View.

Return: tag

or

Return: tag, xy

Types:

problem: Problem

pathName: String

field: Integer

points: Integer

step: Integer

plot: Boolean

name: String

visible: Boolean

tag: Integer

xy: Tuples{Vector{Float64},Vector{Float64}}

LowLevelFEM.showDoFResults — Method

FEM.showDoFResults(problem, q, comp; t=..., name=..., visible=...)

Loads nodal results into a View in gmsh. q is the field to show, comp is the component of the field ("uvec", "ux", "uy", "uz", "vvec", "vx", "vy", "vz"), t is a vector of time steps (same number of columns as q), name is a title to display and visible is a true or false value to toggle on or off the initial visibility in gmsh. If q has more columns, then a sequence of results will be shown (eg. as an animation). This function returns the tag of View.

Return: tag

Types:

problem: Problem

q: Vector{Matrix}

comp: String

t: Vector{Float64}

name: String

visible: Boolean

tag: Integer

LowLevelFEM.showStrainResults — Method

FEM.showStrainResults(problem, E, comp; t=..., name=..., visible=..., smooth=...)

Loads strain results into a View in gmsh. E is a strain field to show, comp is the component of the field ("e", "ex", "ey", "ez", "exy", "eyz", "ezx"), t is a vector of time steps (same length as the number of stress states), name is a title to display, visible is a true or false value to toggle on or off the initial visibility in gmsh and smooth is a true of false value to toggle smoothing the stress field on or off. If length of t is more than one, then a sequence of results will be shown (eg. as an animation). This function returns the tag of View.

Return: tag

Types:

problem: Problem

E: StressField

comp: String

t: Vector{Float64}

name: String

visible: Boolean

smooth: Boolean

tag: Integer

LowLevelFEM.showStressResults — Method

FEM.showStressResults(problem, S, comp; t=..., name=..., visible=..., smooth=...)

Loads stress results into a View in gmsh. S is a stress field to show, comp is the component of the field ("s", "sx", "sy", "sz", "sxy", "syz", "szx", "seqv"), t is a vector of time steps (same length as the number of stress states), name is a title to display, visible is a true or false value to toggle on or off the initial visibility in gmsh and smooth is a true of false value to toggle smoothing the stress field on or off. If length of t is more than one, then a sequence of results will be shown (eg. as an animation). This function returns the tag of View.

Return: tag

Types:

problem: Problem

S: StressField

comp: String

t: Vector{Float64}

name: String

visible: Boolean

smooth: Boolean

tag: Integer

LowLevelFEM.smallestPeriodTime — Method

FEM.smallestPeriodTime(K, M)

Solves the smallest period of time for a dynamic problem given by stiffness matrix K and the mass matrix M.`

Return: Δt

Types:

K: SparseMatrix

M: SparseMatrix

Δt: Float64

LowLevelFEM.solveDisplacement — Method

FEM.solveDisplacement(K, q)

Solves the equation K*q=f for the displacement vector q. K is the stiffness Matrix, q is the load vector.

Return: q

Types:

K: SparseMatrix

f: Vector{Float64}

q: Vector{Float64}

LowLevelFEM.solveStrain — Method

FEM.solveStrain(problem, q)

Solves the strain field E from displacement vector q. Strain field is given per elements, so it usually contains jumps at the boundary of elements. Details of mesh is available in problem.

Return: E

Types:

problem: Problem

q: Vector{Float64}

E: StressField

LowLevelFEM.solveStress — Method

FEM.solveStress(problem, q)

Solves the stress field S from displacement vector q. Stress field is given per elements, so it usually contains jumps at the boundary of elements. Details of mesh is available in problem.

Return: S

Types:

problem: Problem

q: Vector{Float64}

S: StressField

LowLevelFEM.stiffnessMatrix — Method

FEM.stiffnessMatrix(problem)

Solves the stiffness matrix of the problem.

Return: stiffMat

Types:

problem: Problem

stiffMat: SparseMatrix

Index

LowLevelFEM.Problem

LowLevelFEM.StressField

LowLevelFEM.CDM

LowLevelFEM.CDMaccuracyAnalysis

LowLevelFEM.HHT

LowLevelFEM.HHTaccuracyAnalysis

LowLevelFEM.applyBoundaryConditions

LowLevelFEM.applyBoundaryConditions!

LowLevelFEM.applyBoundaryConditions!

LowLevelFEM.dampingMatrix

LowLevelFEM.displacementConstraint

LowLevelFEM.generateMesh

LowLevelFEM.getTagForPhysicalName

LowLevelFEM.initialDisplacement!

LowLevelFEM.initialVelocity!

LowLevelFEM.largestPeriodTime

LowLevelFEM.load

LowLevelFEM.loadVector

LowLevelFEM.massMatrix

LowLevelFEM.material

LowLevelFEM.nodalAcceleration!

LowLevelFEM.nodalForce!

LowLevelFEM.plotOnPath

LowLevelFEM.showDoFResults

LowLevelFEM.showStrainResults

LowLevelFEM.showStressResults

LowLevelFEM.smallestPeriodTime

LowLevelFEM.solveDisplacement

LowLevelFEM.solveStrain

LowLevelFEM.solveStress

LowLevelFEM.stiffnessMatrix

4Serfőző, D., Pere, B.: An effective reduction method with Caughey damping for spurious oscillations in dynamic problems, preprint, https://doi.org/10.21203/rs.3.rs-3930320/v1

1Hilber, Hans M., Thomas JR Hughes, and Robert L. Taylor. "Improved numerical dissipation for time integration algorithms in structural dynamics." Earthquake Engineering & Structural Dynamics 5.3 (1977): 283-292.

2Belytschko, Ted, and Thomas JR, Hughes: "Computational methods for transient analysis", North-Holland, (1983).

3Serfőző, D., Pere, B.: A method to accurately define arbitrary algorithmic damping character as viscous damping. Arch Appl Mech 93, 3581–3595 (2023). https://doi.org/10.1007/s00419-023-02454-9