General API

LowLevelFEM.CoordinateSystem — Type

CoordinateSystem(vec1, vec2)

A structure containing the data of a coordinate system.

vec1: direction of the new x axis.

vec2: together with vec1 determine the xy plane

If the problem is two dimensional, it is enough to give the first two elements of vec1. Elements of vec1 and vec2 can be functions. In 3D case the functions have three arguments (x, y, and z coordinates), otherwise (in 2D case) the number of arguments is two (x and y coordinates).

Types:

vec1: Vector{Float64}

vec2: Vector{Float64}

Examples

# 2D case
nx(x, y) = x
ny(x, y) = y
cs = CoordinateSystem([nx, ny])
Q = rotateNodes(problem, "body", cs)
q2 = Q' * q1 # where `q1` is in Cartesian, `q2` is in Axisymmetric coordinate system and
# `q1` is a nodal displacement vector.
S2 = Q' * S1 * Q # where `S1` is a stress field in Cartesian coordinate system while
# `S2` is in Axisymmetric coordinate system.

# 3D case
n1x(x, y, z) = x
n1y(x, y, z) = y
n2x(x, y, z) = -y
n2y = n1x
cs = CoordinateSystem([n1x, n1y, 0], [n2x, n2y, 0])
Q = rotateNodes(problem, "body", cs)

LowLevelFEM.Eigen — Type

Eigen(f, ϕ, model)

A structure containing the eigenfrequencies and eigen modes.

f: eigenfrequencies

ϕ: eigen modes

model: same as Problem

Types:

f: Matrix{Float64}

ϕ: Vector{Float64}

model: Problem

LowLevelFEM.Material — Type

Material(phName, type, E, ν, ρ, k, c, α, λ, μ, κ)

Structure containing the material type and constants.

type: constitutive law (:Hooke, :StVenantKirchhoff, :NeoHookeCompressible)

E: elastic modulus

ν: Poisson's ratio

ρ: mass density

k: thermal conductivity

c: specific heat

α: thermal expansion coefficient

λ: Lamé parameter

μ: Lamé parameter

κ: bulk modulus

phName is the name of the physical group where the material is used.

Types:

phName: String

type: Symbol

E: Float64

ν: Float64

ρ: Float64

k: Float64

c: Float64

α: Float64

λ: Float64

μ: Float64

κ: Float64

LowLevelFEM.Problem — Type

Problem(materials; thickness=..., type=..., bandwidth=...)

Structure containing key data for a problem.

Parts of the model with their material constants. Multiple materials can be provided (see material).

Problem type: :Solid, :PlaneStrain, :PlaneStress, :AxiSymmetric, :HeatConduction, :PlaneHeatConduction,

:AxiSymmetricHeatConduction, :Truss. For :AxiSymmetric, the symmetry axis is y, and the geometry must be drawn in the positive x half-plane.

Bandwidth optimization using Gmsh's built-in reordering. Options: :RCMK, :Hilbert, :Metis, or :none (default).

Dimension of the problem, determined by type.

Material constants (vector of Material; see the Material struct).

Plate thickness (for 2D plate problems).

Number of nodes (non).

Geometry dimension.

Problem dimension (e.g., a 3D heat conduction problem is a 1D problem).

In case of 2D truss displacements have to be fixed in the third direction.

Types:

materials: Material

type: Symbol

bandwidth: Symbol

dim: Integer

thickness: Float64

non: Integer

dim: Integer

pdim: Integer

LowLevelFEM.ScalarField — Type

ScalarField(A, a, t, numElem, nsteps, type, model)
ScalarField(problem, dataField)

Structure containing all data of a scalar field (e.g., temperature).

A: vector of element-wise scalar data

a: matrix of nodal values of the scalar field

numElem: vector of element tags

nsteps: number of time steps stored in A (for animations)

type: type of data (e.g., :T for temperature)

model: associated Problem

Types:

A: Vector{Vector{Float64}}

a: Matrix{Float64}

t: Vector{Float64}

numElem: Vector{Integer}

nsteps: Integer

type: Symbol

model: Problem

Example

s(x,y,z) = 2x + 3y
fs = field("body", f=s)
S = ScalarField(problem, [fs])

Here S is defined element-wise.

LowLevelFEM.SystemMatrix — Type

SystemMatrix(A::SparseMatrixCSC{Float64}, model::Problem)

Structure containing the stiffness/mass/heat conduction/heat capacity/latent heat/... matrix and the associated Problem.

Types:

A: SparseMatrixCSC{Float64}

model: Problem

LowLevelFEM.TensorField — Type

TensorField(A, a, t, numElem, nsteps, type, model)
TensorField(problem, dataField)

Structure containing the data of a tensor field (e.g., stress or strain).

A: tensor of element-wise tensor data

a: matrix of nodal values of the tensor field

numElem: vector of element tags

nsteps: number of time steps stored in A (for animations)

type: type of data (e.g., :u, :q)

model: associated Problem

Types:

A: Vector{Matrix{Float64}}

a: Matrix{Float64}

t: Vector{Float64}

numElem: Vector{Integer}

nsteps: Integer

type: Symbol

model: Problem

Example

tx(x,y,z) = x + y
txy(x,y,z) = z
ft = field("body", fx=tx, fxy=txy, fz=3)
T = TensorField(problem, [ft])

Here T is defined element-wise

LowLevelFEM.Transformation — Type

Transformation(T::SparseMatrixCSC{Float64}, non::Int64, dim::Int64)

Structure containing the transformation matrix T at each node in the FE mesh, the number of nodes non, and the problem dimension dim.

Types:

T: SparseMatrixCSC{Float64}

non: Int64

dim: Int64

LowLevelFEM.VectorField — Type

VectorField(A, a, t, numElem, nsteps, type, model)
VectorField(problem, dataField)

Structure containing the data of a vector field (e.g., displacement or heat flux).

A: vector of element-wise vector data

a: matrix of nodal values of the vector field

numElem: vector of element tags

nsteps: number of time steps stored in A (for animations)

type: type of data (e.g., :u, :q)

model: associated Problem

Types:

A: Vector{Matrix{Float64}}

a: Matrix{Float64}

t: Vector{Float64}

numElem: Vector{Integer}

nsteps: Integer

type: Symbol

model: Problem

Example

vx(x,y,z) = x + y
vy(x,y,z) = z
fv = field("body", fx=vx, fy=vy, fz=3)
V = VectorField(problem, [fv])

Here V is defined element-wise

Base.copy — Method

Base.copy(A::SystemMatrix)

Internal function to copy the whole content of a SystemMatrix.

Base.show — Method

Base.show(io::IO, M::SystemMatrix)

Internal function to display SystemMatrix as a SparseMatrixCSC{Float64}.

Base.show — Method

Base.show(io::IO, M::Union{ScalarField,VectorField,TensorField})

Internal function to display ScalarField, VectorField and TensorField as a Matrix{Float64} or Vector{Matrix{Float64}}.

LowLevelFEM.constrainedDoFs — Method

constrainedDoFs(problem, supports)

Returns the serial numbers of constrained degrees of freedom. Support is a vector of boundary conditions given with the function displacementConstraint.

Return: DoFs

Types:

problem: Problem

supports: Vector{Tuple{String, Float64, Float64, Float64}}

DoFs: Vector{Int64}

LowLevelFEM.displacementConstraint — Method

displacementConstraint(name; ux=..., uy=..., uz=...)

Specifies displacement constraints on the physical group name. At least one of ux, uy, or uz must be provided (depending on the problem dimension). ux, uy, or uz can be a constant or a function of x, y, and z. (e.g., fn(x,y,z) = 5*(5-x); displacementConstraint("support1", ux=fn))

Returns: Tuple{String, Float64 or Function, Float64 or Function, Float64 or Function}

Examples

hc = heatConvection("outer", h=10.0, Tₐ=20.0)

Examples

src = heatSource("body", h=1.0e6)

Examples

q = heatFlux("out", qn=500.0)

Examples

bcT = temperatureConstraint("hot_face", T=100.0)

Examples

ld = load("load", fy=-1.0)

Examples

bc = displacementConstraint("supp", ux=0, uy=0)

Types:

name: String

ux: Float64 or Function

uy: Float64 or Function

uz: Float64 or Function

LowLevelFEM.elasticSupport — Method

elasticSupport(name; kx=..., ky=..., kz=...)

Specifies distributed stiffness for an elastic support on the physical group name. kx, ky, or kz can be a constant or a function of x, y, and z. (e.g., fn(x,y,z) = 5*(5-x); elasticSupport("supp1", kx=fn)) Default values are 1.

Returns: Tuple{String, Float64 or Function, Float64 or Function, Float64 or Function}

Types:

name: String

kx: Float64 or Function

ky: Float64 or Function

kz: Float64 or Function

LowLevelFEM.elementsToNodes — Method

elementsToNodes(T)

Solves the nodal results F from the elemental results T. T can be ScalarField, VectorField or TensorField.

Return: F

Types:

T: ScalarField, VectorField or TensorField

F: ScalarField, VectorField or TensorField

LowLevelFEM.expandTo3D — Method

expandTo3D(v2D::VectorField)

Expand a 2D vector field into 3D by adding a zero z-component.

return: VectorField

Examples

V3D = expandTo3D(V2D)

LowLevelFEM.field — Method

field(name; f=..., fx=..., fy=..., fz=..., fxy=..., fyz=..., fzx=...)

Specifies the value of a scalar, vector, or tensor field on the physical group name. At least one of fx, fy, or fz etc. must be provided (depending on the problem dimension). Each component can be a constant or a function of x, y, and z. (e.g., fn(x,y,z) = 5*(5-x); field("surf1", fx=fn))

Returns: Tuple{String, Float64 or Function, Float64 or Function, Float64 or Function, ... x7}

Types:

name: String

f: Float64 or Function

fx: Float64 or Function

fy: Float64 or Function

fz: Float64 or Function

fxy: Float64 or Function

fyz: Float64 or Function

fzx: Float64 or Function

Examples

f1(x, y, z) = y
f2 = field("face1", f=f1)
qq = scalarField(problem, [f2])
qqq = showDoFResults(qq, :scalar)

LowLevelFEM.fieldError — Method

fieldError(F)

Computes the deviation of field results F (stresses, strains, heat flux components) at nodes where the field has jumps. The result can be displayed with the showDoFResults function.

Returns: e

Types:

F: VectorField or TensorField

e: ScalarField

LowLevelFEM.freeDoFs — Method

freeDoFs(problem, supports)

Returns the serial numbers of unconstrained degrees of freedom. Support is a vector of boundary conditions given with the function displacementConstraint.

Return: DoFs

Types:

problem: Problem

supports: Vector{Tuple{String, Float64, Float64, Float64}}

DoFs: Vector{Int64}

LowLevelFEM.getEigenValues — Method

getEigenValues(A::VectorField)

A function to extract the elements of a vector field to separate scalar fields.

Return: λ1, λ2, λ3

Types:

A: VectorField

λ1: ScalarField

λ2: ScalarField

λ3: ScalarField

Examples:

using LinearAlgebra
λ, Q = eigen(S)
λ1, λ2, λ2 = getEigenValues(λ)

LowLevelFEM.getEigenVectors — Method

getEigenVectors(A::TensorField)

A function to extract the columns of a tensor field to separate vector fields.

Return: N1, N2, N3

Types:

A: TensorField

N1: VectorField

N2: VectorField

N3: VectorField

Examples:

using LinearAlgebra
λ, Q = eigen(S)
N1, N2, N2 = getEigenVectors(Q)

LowLevelFEM.heatConvection — Method

heatConvection(name; h=..., Tₐ=...)

Specifies convective boundary conditions on the surface in the physical group name. h is the heat transfer coefficient of the surrounding medium; Tₐ is the ambient temperature. Tₐ can be either a constant or a function of x, y, and z.

Returns: Tuple{String, Float64 or Function, Float64 or Function, Float64 or Function}

Types:

name: String

h: Float64

Tₐ: Float64 or Function

LowLevelFEM.heatFlux — Method

heatFlux(name; qn=...)

Specifies the heat flux normal to the surface of the physical group name. qn can be a constant or a function of x, y, and z. (e.g., fn(x,y,z) = 5*(5-x); load("flux1", qn=fn))

Returns: Tuple{String, Float64 or Function, Float64 or Function, Float64 or Function}

Types:

name: String

qn: Float64 or Function

LowLevelFEM.heatSource — Method

heatSource(name; h=...)

Specifies the volumetric heat source in the physical group name. h can be a constant or a function of x, y, and z. (e.g., fn(x,y,z) = 5*(5-x); load("source1", h=fn))

Returns: Tuple{String, Float64 or Function, Float64 or Function, Float64 or Function}

Types:

name: String

h: Float64 or Function

LowLevelFEM.integrate — Method

integrate(problem::Problem, phName::String, f::Union{Function,ScalarField})

Integrates the function or scalar field f over the physical group phName defined in the geometry of problem.

Returns: integral

Types:

problem: Problem

phName: String

f: Function (of x, y and z) or ScalarField

integral: Number

Examples:

f(x, y, z) = x^2 + y^2
Iz = integrate(prob, "body", f)

LowLevelFEM.isElementwise — Method

isElementwise(field)

Check if a given field is defined per element (elementwise quantity).

Elementwise quantities are associated with finite elements as a whole, for example stresses, strains, or energy densities evaluated inside elements.

Examples

isElementwise(displacement_field)   # returns false
isElementwise(strain_field)         # returns true

LowLevelFEM.isNodal — Method

isNodal(field)

Check if a given field is defined at nodes (nodal quantity).

Nodal quantities are associated with mesh nodes, for example displacements, nodal forces, or nodal temperatures.

Examples

isNodal(displacement_field)   # returns true
isNodal(strain_field)         # returns false

LowLevelFEM.isSaved — Method

isSaved(fileName::String)

Checks whether a variable has been saved or not.

Returns: Boolean

Types:

fileName: String

LowLevelFEM.load — Method

load(name; fx=..., fy=..., fz=...)

Specifies a distributed load on the physical group name. At least one of fx, fy, or fz must be provided (depending on the problem dimension). fx, fy, or fz can be a constant or a function of x, y, and z or ScalarField. (e.g., fn(x,y,z) = 5*(5-x); load("load1", fx=fn))

Returns: Tuple{String, Float64 or Function, Float64 or Function, Float64 or Function}

Types:

name: String

fx: Float64 or Function

fy: Float64 or Function

fz: Float64 or Function

LowLevelFEM.loadField — Method

loadField(fileName::String)

Loads a ScalarField, VectorField, or TensorField from a file named fileName (without "-LLF-Data.jld2").

Returns: variable

Types:

fileName: String

variable: ScalarField, VectorField or TensorField

LowLevelFEM.material — Method

material(name; type=:Hooke, E=2.0e5, ν=0.3, ρ=7.85e-9, k=45, c=4.2e8, α=1.2e-5, λ=νE/(1+ν)/(1-2ν), μ=E/(1+ν)/2, κ=E/(1-2ν)/3)

Returns a structure in which name is the name of a physical group, type is the name of the constitutive law (e.g., :Hooke), E is the modulus of elasticity, ν is Poisson's ratio, and ρ is the mass density. k is the thermal conductivity, c is the specific heat, α is the coefficient of thermal expansion, λ and μ are the Lamé parameters, and κ is the bulk modulus.

Returns: mat

Examples

mat = material("body", E=210e3, ν=0.3, ρ=7.85e-9)

Types:

mat: Material

name: String

type: Symbol

E: Float64

ν: Float64

ρ: Float64

k: Float64

c: Float64

α: Float64

λ: Float64

μ: Float64

κ: Float64

LowLevelFEM.nodesToElements — Method

nodesToElements(T, onPhysicalGroup="")

Solves the element results F from the nodal results T. T can be ScalarField, VectorField or TensorField. If onPhysicalGroup is an existing physical group in gmsh, field T will be solve only on elements belonging to that physical group. Dimension of physical group can be different than the dimension of the problem. (eg. mapping from 3D volume to a 2D surface)

Return: F

Types:

T: ScalarField, VectorField or TensorField

F: ScalarField, VectorField or TensorField

onPhysicalGroup: String

LowLevelFEM.normalVector — Method

normalVector(problem::Problem, phName::String) -> VectorField

Compute outward unit normal vectors for all nodes of a surface-type physical group in a 3D Gmsh model.

Arguments

problem::Problem: A Problem structure containing the current Gmsh model (name, dimension, number of nodes, etc.).

phName::String: The name of a physical surface group in Gmsh for which the normal vectors are computed.

Description

The function sets the current model, queries the elements and nodes that belong to the given physical surface group, and evaluates the gradients of the Lagrange basis functions to compute local tangent vectors of the surface. Normal vectors are obtained as cross products of these tangents and normalized to unit length.

Each node belonging to the physical surface group is assigned its corresponding 3D unit normal vector.

Returns

VectorField: A VectorField structure that stores the nodal normal vectors on the given physical surface.

Errors

Throws an error if the provided physical group is not of surface type (dim != 2).

Example

using LowLevelFEM

# Load a 3D geometry and mesh it in Gmsh
gmsh.initialize()
gmsh.open("box_with_surface.msh")

# Define a 3D model on the volume physical group "body"
mat = material("body")
prob = Problem([mat])

# Compute nodal normals on a physical surface named "leftWall"
nv = normalVector(problem, "leftWall")

# Show the normal vectors on the model
showDoFResults(nv)
openPostProcessor()

LowLevelFEM.openPostProcessor — Method

openPostProcessor(; model=...)

Launches the Gmsh postprocessor window with the postprocessor tree opened (of model).

Returns: nothing

Types:

model: Int64

LowLevelFEM.openPreProcessor — Method

openPreProcessor(; openGL=...)

Launches the Gmsh preprocessor window with OpenGL disabled by default.

Returns: nothing

Types:

openGL: Boolean

LowLevelFEM.plotOnPath — Method

plotOnPath(problem, pathName, field; points=100, step=..., plot=..., name=..., visible=..., offsetX=..., offsetY=..., offsetZ=...)

Loads a 2D plot along a path into a View in Gmsh. field is the View id in Gmsh from which the field data is imported. pathName is the name of a physical group that contains a curve. The curve is divided into equal-length segments with points sampling points. The field is shown at these points. step is the sequence number of the displayed step. If no step is given, it shows all available steps as an animation. If plot is true, an additional return parameter (a tuple of vectors) is returned, where x is the horizontal axis and y is the vertical axis of the plot (see the Plots package). name is the title of the graph, and visible is a Boolean to toggle the initial visibility in Gmsh on or off. This function returns the tag of the View.

Returns: tag

or

Returns: tag, xy

Types:

problem: Problem

pathName: String

field: Integer

points: Integer

step: Integer

plot: Boolean

name: String

visible: Boolean

tag: Integer

xy: Tuples{Vector{Float64},Vector{Float64}}

LowLevelFEM.probe — Method

probe(A::Union{ScalarField,VectorField,TensorField}, x::Number, y::Number, z::Number; step=Int)

Get the value of the field A at point coordinates x, y, z at time step step.

Returns: Float64 or Vector{Float64} or Matrix{Float64}

Types:

A: ScalarField or VectorField or TensorField

x: Number

y: Number

z: Number

step: Int

LowLevelFEM.probe — Method

probe(A::Union{ScalarField,VectorField,TensorField}, s::String; step=Int)

Get the value of the field A at a point given by its physical name in Gmsh at time step step.

Returns: Float64 or Vector{Float64} or Matrix{Float64}

Types:

A: ScalarField or VectorField or TensorField

x: Number

y: Number

z: Number

step: Int

LowLevelFEM.projectTo2D — Method

projectTo2D(v3D::VectorField)

Project a 3D vector field onto the xy-plane by discarding the z-component.

return: VectorField

Examples

V2D = expandTo3D(V3D)

LowLevelFEM.resultant — Method

resultant(problem, field, phName; grad=false, component=:x)

Computes the resultant of field on the physical group phName. Returns the resultant(s) in a tuple. The number of elements in the tuple depends on the dimension of problem. It can solve the resultant of a load vector (sum of the elements of the vector), if field is a vector of floats. If field is a view (tag of a view in gmsh), then the integral of the field is solved. field must have only one component. If grad is true, then the gradient of the field will be evaluated and component of the gradient (:x, :y or :z) will be used to solve the resultant.

Return: res

or

Return: resx, resy

or

Return: resx, resy, resz

Types:

field: Vector{Float64}

phName: String

dim: Int64

res: Float64

resx: Float64

resy: Float64

resz: Float64

LowLevelFEM.rotateNodes — Method

rotateNodes(problem, phName, CoordSys)

Creates the T transformation matrix, which rotates the nodal coordinate system of the nodes in phName physical group to the coordinate systen defined by CoordSys. The mesh belongs to problem.

Return: T

Types:

problem: Problem

phName: String

CoordSys: CoordinateSystem

T: SparseMatrix

LowLevelFEM.saveField — Method

saveField(fileName::String, variable::Union{ScalarField,VectorField,TensorField,Number})

Saves variable of type ScalarField, VectorField, or TensorField to a file named fileName. The name of the file will be complemented with the string "-LLF-Data.jld2"

Returns: nothing

Types:

fileName: String

variable: ScalarField, VectorField or TensorField

LowLevelFEM.scalarField — Method

scalarField(problem, dataField)

Defines a scalar field from dataField, which is a tuple of name of physical group and prescribed values or functions. Mesh details are in problem.

Return: Vector{Float64}

Types:

problem: Problem

dataField: Vector{Tuple{String, Float64,...}}

Examples

f2 = field("face1", f=1)
qq = scalarField(problem, [f2])
qqq = showDoFResults(qq, :scalar)

Here ScalarField is defined in nodes.

LowLevelFEM.setParameter — Method

setParameter(name, value)

Defines a parameter name and sets its value to value.

Returns: nothing

Types:

name: String

value: Float64

LowLevelFEM.setParameters — Method

setParameters(name, value)

Defines a parameter name and sets its value to value, which is a Vector{Float64}.

Returns: nothing

Types:

name: String

value: Vector{Float64}

LowLevelFEM.showBucklingResults — Method

showBucklingResults(Φ, name=..., visible=...)

Loads buckling results into a View in Gmsh. Φ is an Eigen struct. name is a title to display and visible is a Boolean to toggle the initial visibility in Gmsh on or off. Click on ▷| to change the results. This function returns the tag of the View.

Returns: tag

Types:

Φ: Eigen

name: String

visible: Boolean

tag: Integer

LowLevelFEM.showDoFResults — Method

showDoFResults(q, comp; name=..., visible=...)

Loads nodal results into a View in Gmsh. q is the field to show, comp is the component of the field (:vector, :uvec, :ux, :uy, :uz, :vvec, :vx, :vy, :vz, :qvec, :qx, :qy, :qz, :T, :p, :qn, :s, :sx, :sy, :sz, :sxy, :syx, :syz, :szy, :szx, :sxz, :e, :ex, :ey, :ez, :exy, :eyx, :eyz, :ezy, :ezx, :exz, :seqv, :scalar, :tensor), name is a title to display and visible is a Boolean to toggle the initial visibility in Gmsh on or off. If q has more columns, then a sequence of results will be shown (e.g., as an animation). This function returns the tag of the View.

Returns: tag

Types:

q: ScalarField, VectorField or TensorField

comp: Symbol

name: String

visible: Boolean

tag: Integer

LowLevelFEM.showElementResults — Method

showElementResults(F, comp; name=..., visible=..., smooth=...)

Same as ShowStressResults or showStrainResults, depending on the type of F data field.

Return: tag

Types:

F: TensorField

comp: Symbol

name: String

visible: Boolean

smooth: Boolean

tag: Integer

LowLevelFEM.showHeatFluxResults — Method

showHeatFluxResults(Q, comp; name=..., visible=..., smooth=...)

Loads heat flux results into a View in Gmsh. Q is a heat flux field to show, comp is the component of the field (:qvec, :qx, :qy, :qz, :q), name is a title to display, visible is a Boolean to toggle the initial visibility in Gmsh on or off, and smooth is a Boolean to toggle smoothing on or off. If Q contains more than one time step, a sequence of results will be shown (e.g., as an animation). This function returns the tag of the View.

Returns: tag

Types:

S: VectorField

comp: Symbol

name: String

visible: Boolean

smooth: Boolean

tag: Integer

LowLevelFEM.showModalResults — Method

showModalResults(Φ, name=..., visible=...)

Loads modal results into a View in Gmsh. Φ is an Eigen struct. name is a title to display and visible is a Boolean to toggle the initial visibility in Gmsh on or off. Click on ▷| to change the results. This function returns the tag of the View.

Returns: tag

Types:

Φ: Eigen

name: String

visible: Boolean

tag: Integer

LowLevelFEM.showOnSurface — Method

showOnSurface(field, phName; grad=false, component=:x, offsetX=0, offsetY=0, offsetZ=0, name=phName, visible=false)

Shows the values of a scalar field on a surface with physical name phName. field is the tag of a View in Gmsh. The values of the field are calculated at the intersection with the surface. grad is a Boolean to toggle the gradient of the field on or off. component is the component of the gradient of field (:x, :y, :z) to be shown. offsetX, offsetY, offsetZ are the offsets in the x, y, and z directions where the values are sampled. name is a title to display, and visible is a Boolean to toggle the initial visibility in Gmsh on or off.

Returns: tag

Types:

field: Integer

phName: String

grad: Boolean

component: Symbol

offsetX: Float64

offsetY: Float64

offsetZ: Float64

name: String

visible: Boolean

tag: Integer

LowLevelFEM.showStrainResults — Method

showStrainResults(E, comp; name=..., visible=..., smooth=...)

Loads strain results into a View in Gmsh. E is a strain field to show, comp is the component of the field (:e, :ex, :ey, :ez, :exy, :eyz, :ezx), name is a title to display, visible is a Boolean to toggle the initial visibility in Gmsh on or off and smooth is a Boolean to toggle smoothing the stress field on or off. If E contains more than one time steps, then a sequence of results will be shown (e.g., as an animation). This function returns the tag of the View.

Returns: tag

Types:

E: TensorField

comp: Symbol

name: String

visible: Boolean

smooth: Boolean

tag: Integer

LowLevelFEM.showStressResults — Method

showStressResults(S, comp; name=..., visible=..., smooth=...)

Loads stress results into a View in Gmsh. S is a stress field to show, comp is the component of the field (:s, :sx, :sy, :sz, :sxy, :syz, :szx, :seqv), name is a title to display, visible is a Boolean to toggle the initial visibility in Gmsh on or off, and smooth is a Boolean to toggle smoothing the stress field on or off. If S contains more than one time step, a sequence of results will be shown (e.g., as an animation). This function returns the tag of the View.

Returns: tag

Types:

S: TensorField

comp: Symbol

name: String

visible: Boolean

smooth: Boolean

tag: Integer

LowLevelFEM.temperatureConstraint — Method

temperatureConstraint(name; T=...)

Specifies temperature constraints on the physical group name. T can be a constant or a function of x, y, and z. (e.g., fn(x,y,z) = 5*(5-x); temperatureConstraint("surf1", T=fn))

Returns: Tuple{String, Float64 or Function, Float64 or Function, Float64 or Function}

Types:

name: String

T: Float64 or Function

LowLevelFEM.tensorField — Method

tensorField(problem, dataField; type=...)

Defines a vector field from dataField, which is a tuple of name of physical group and prescribed values or functions. Mesh details are in problem. type can be an arbitrary Symbol, e.g., :u or :f.

Return: TensorField

Types:

problem: Problem

dataField: Vector{Tuple{String, Float64,...}}

Examples

f1(x, y, z) = sin(x)
f2(x, y, z) = 5y
ff1 = field("face1", fx=f1, fy=f2, fz=0, fxy=1, fyz=1, fzx=f2)
ff2 = field("face2", fx=f2, fy=f1, fz=1, fxy=1, fyz=f1, fzx=1)
qq = tensorField(problem, [ff1, ff2])
qq0 = showDoFResults(qq, :tensor)

Here TensorField is defined in nodes.

LowLevelFEM.vectorField — Method

vectorField(problem, dataField; type=...)

Defines a vector field from dataField, which is a tuple of name of physical group and prescribed values or functions. Mesh details are in problem. type can be an arbitrary Symbol, e.g., :u or :f.

Return: VectorField

Types:

problem: Problem

dataField: Vector{Tuple{String, Float64,...}}

Examples

f1(x, y, z) = sin(x)
f2(x, y, z) = 5y
ff1 = field("face1", fx=f1, fy=f2, fz=0)
ff2 = field("face2", fx=f2, fy=f1, fz=1)
qq = vectorField(problem, [ff1, ff2])
qq0 = showDoFResults(qq, :vector)

Here VectorField is defined in nodes.